fix(coin-request): fixed float type to double for better match

+++ 2025_PhyoNyiNyiPaing/src/main/java/com/paulphyo/api/CoinRequest.java
@@ -19,14 +19,14 @@ import java.util.List;
  * </pre>
  */
 public class CoinRequest {
-    private float targetAmount;
+    private double targetAmount;
     private List<Double> coinDenominations;
 
-    public float getTargetAmount() {
+    public double getTargetAmount() {
         return targetAmount;
     }
 
-    public void setTargetAmount(float targetAmount) {
+    public void setTargetAmount(double targetAmount) {
         this.targetAmount = targetAmount;
     }
